Some bugfixes for geostrophic_interpolation/coriolis/mesh option

+++ assemble/Geostrophic_Pressure.F90
@@ -2219,7 +2219,7 @@ contains
     type(mesh_type), pointer :: new_p_mesh, new_u_mesh, old_p_mesh, old_u_mesh
     type(scalar_field) :: new_p, old_w, old_p, new_w
     type(vector_field) :: coriolis, conserv, new_res, old_res
-    type(vector_field), pointer :: lold_velocity, new_positions, old_positions
+    type(vector_field), pointer :: old_bc_velocity, new_positions, old_positions
 
     logical :: gp
     character(len = OPTION_PATH_LEN) :: gp_mesh_name
@@ -2257,17 +2257,17 @@ contains
     if(stat == SPUD_NO_ERROR) then
       old_u_mesh => extract_mesh(old_state, u_mesh_name)
       new_u_mesh => extract_mesh(new_state, u_mesh_name)
-      allocate(lold_velocity)
-      call allocate(lold_velocity, old_velocity%dim, old_u_mesh, old_velocity%name)
-      lold_velocity%option_path = old_velocity%option_path
-      call remap_field(old_velocity, lold_velocity)
-      call populate_vector_boundary_conditions(lold_velocity, trim(complete_field_path(lold_velocity%option_path)) // "/boundary_conditions", old_positions)
+      allocate(old_bc_velocity)
+      call allocate(old_bc_velocity, old_velocity%dim, old_u_mesh, old_velocity%name)
+      old_bc_velocity%option_path = old_velocity%option_path
+      call zero(old_bc_velocity)
+      call populate_vector_boundary_conditions(old_bc_velocity, trim(complete_field_path(old_bc_velocity%option_path)) // "/boundary_conditions", old_positions)
     else
       old_u_mesh => old_velocity%mesh
       new_u_mesh => new_velocity%mesh
-      allocate(lold_velocity)
-      lold_velocity = old_velocity
-      call incref(lold_velocity)
+      allocate(old_bc_velocity)
+      old_bc_velocity = old_velocity
+      call incref(old_bc_velocity)
     end if
     call get_option(trim(base_path) // "/conservative_potential/mesh/name", p_mesh_name)
     old_p_mesh => extract_mesh(old_state, p_mesh_name)
@@ -2276,8 +2276,8 @@ contains
     ! Compute the old Coriolis
                 
     call allocate(coriolis, dim, old_u_mesh, "Coriolis")
-    coriolis%option_path = lold_velocity%option_path
-    call coriolis_from_velocity(old_state, lold_velocity, coriolis)
+    coriolis%option_path = old_velocity%option_path
+    call coriolis_from_velocity(old_state, old_velocity, coriolis)
         
     call allocate(old_p, old_p_mesh, gi_conservative_potential_name)
     old_p%option_path = trim(base_path) // "/conservative_potential"
@@ -2328,35 +2328,50 @@ contains
     call allocate(conserv, dim, old_u_mesh, name = "CoriolisConservative")
     if(gp) then
       call geopressure_decomposition(old_state, coriolis, old_gp)
-      call projection_decomposition(old_state, coriolis, old_p, bcfield = lold_velocity, matrices = matrices, gp = old_gp) 
+      call projection_decomposition(old_state, coriolis, old_p, bcfield = old_bc_velocity, matrices = matrices, gp = old_gp) 
       call correct_velocity(matrices, old_res, old_p, conserv = conserv, gp = old_gp)
     else
-      call projection_decomposition(old_state, coriolis, old_p, bcfield = lold_velocity, matrices = matrices) 
+      call projection_decomposition(old_state, coriolis, old_p, bcfield = old_bc_velocity, matrices = matrices) 
       call correct_velocity(matrices, old_res, old_p, conserv = conserv)
     end if
     
     if(debug_vtus) then
-      call allocate(div, old_p_mesh, trim(lold_velocity%name) // "Divergence")
-      call set_solver_options(temp_solver_path, ksptype = "cg", pctype = "sor", rtol = 0.0, atol = epsilon(0.0), max_its = 10000)
-      div%option_path = temp_solver_path
-      call zero(div)
-      call compute_divergence(lold_velocity, matrices%ct_m, get_mass_matrix(old_state, old_p_mesh), div)  
-      if(gp) then
-        call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
-          & sfields = (/old_p, old_gp, div/), vfields = (/lold_velocity, coriolis, conserv, old_res/), stat = stat)
+      if(old_velocity%mesh == old_u_mesh) then
+        call allocate(div, old_p_mesh, trim(old_velocity%name) // "Divergence")
+        call set_solver_options(temp_solver_path, ksptype = "cg", pctype = "sor", rtol = 0.0, atol = epsilon(0.0), max_its = 10000)
+        div%option_path = temp_solver_path
+        call zero(div)
+        call compute_divergence(old_velocity, matrices%ct_m, get_mass_matrix(old_state, old_p_mesh), div)  
+        if(gp) then
+          call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
+            & sfields = (/old_p, old_gp, div/), vfields = (/old_velocity, coriolis, conserv, old_res/), stat = stat)
+        else
+          call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
+            & sfields = (/old_p, div/), vfields = (/old_velocity, coriolis, conserv, old_res/), stat = stat)
+        end if
+        if(stat /= 0) then
+          ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
+        end if
+        call delete_option(div%option_path)
+        call deallocate(div)
       else
-        call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
-          & sfields = (/old_p, div/), vfields = (/lold_velocity, coriolis, conserv, old_res/), stat = stat)
+        if(gp) then
+          call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
+            & sfields = (/old_p, old_gp/), vfields = (/old_velocity, coriolis, conserv, old_res/), stat = stat)
+        else
+          call vtk_write_fields("geostrophic_interpolation_old", vtu_index, old_positions, model = old_u_mesh, &
+            & sfields = (/old_p/), vfields = (/old_velocity, coriolis, conserv, old_res/), stat = stat)
+        end if
+        if(stat /= 0) then
+          ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
+        end if
       end if
-      if(stat /= 0) then
-        ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
-      end if
-      call delete_option(div%option_path)
-      call deallocate(div)
     end if
     call deallocate(conserv)
         
     call deallocate(coriolis)
+    call deallocate(old_bc_velocity)
+    deallocate(old_bc_velocity)
     
     ! Insert the horizontal Velocity residual    
     
@@ -2369,7 +2384,7 @@ contains
       ! Insert the vertical Velocity
       call allocate(old_w, old_u_mesh, gi_w_name)
       old_w%option_path = old_res%option_path
-      call set(old_w, lold_velocity, W_)
+      call set(old_w, old_velocity, W_)
       ewrite_minmax(old_w%val)
       
       call allocate(new_w, new_u_mesh, old_w%name)
@@ -2477,8 +2492,6 @@ contains
     end if
     
     call deallocate(matrices)
-    call deallocate(lold_velocity)
-    deallocate(lold_velocity)
     
     if(debug_vtus) then
       vtu_index = vtu_index + 1
@@ -2564,7 +2577,7 @@ contains
     type(scalar_field) :: res_p
     type(scalar_field) :: new_p, new_w
     type(vector_field) :: conserv, coriolis, new_res
-    type(vector_field), pointer :: lnew_velocity, new_positions
+    type(vector_field), pointer :: new_bc_velocity, new_positions
 
     logical :: gp
     type(scalar_field) :: new_gp
@@ -2598,16 +2611,16 @@ contains
     call get_option(trim(base_path) // "/coriolis/mesh/name", u_mesh_name, stat = stat)
     if(stat == SPUD_NO_ERROR) then
       new_u_mesh => extract_mesh(new_state, u_mesh_name)
-      allocate(lnew_velocity)
-      call allocate(lnew_velocity, new_velocity%dim, new_u_mesh, new_velocity%name)
-      lnew_velocity%option_path = new_velocity%option_path
-      call remap_field(new_velocity, lnew_velocity)
-      call populate_vector_boundary_conditions(lnew_velocity, trim(complete_field_path(lnew_velocity%option_path)) // "/boundary_conditions", new_positions)
+      allocate(new_bc_velocity)
+      call allocate(new_bc_velocity, new_velocity%dim, new_u_mesh, new_velocity%name)
+      new_bc_velocity%option_path = new_velocity%option_path
+      call zero(new_bc_velocity)
+      call populate_vector_boundary_conditions(new_bc_velocity, trim(complete_field_path(new_bc_velocity%option_path)) // "/boundary_conditions", new_positions)
     else
       new_u_mesh => new_velocity%mesh
-      allocate(lnew_velocity)
-      lnew_velocity = new_velocity
-      call incref(lnew_velocity)
+      allocate(new_bc_velocity)
+      new_bc_velocity = new_velocity
+      call incref(new_bc_velocity)
     end if
     call get_option(trim(base_path) // "/conservative_potential/mesh/name", p_mesh_name)
     new_p_mesh => extract_mesh(new_state, p_mesh_name)
@@ -2654,11 +2667,11 @@ contains
       call allocate(res_p, new_p_mesh, trim(new_res%name) // "ConservativePotential")
       res_p%option_path = new_p%option_path
       call zero(res_p)
-      call projection_decomposition(new_state, new_res, res_p, bcfield = lnew_velocity, matrices = matrices)
+      call projection_decomposition(new_state, new_res, res_p, bcfield = new_bc_velocity, matrices = matrices)
       call correct_velocity(matrices, coriolis, res_p)
       call deallocate(res_p)
     else
-      call allocate(matrices, new_state, new_res, new_p, bcfield = lnew_velocity, add_cmc = .false.)
+      call allocate(matrices, new_state, new_res, new_p, bcfield = new_bc_velocity, add_cmc = .false.)
     end if
 
     ! Add the conservative component  
@@ -2687,33 +2700,46 @@ contains
     
     ! Invert for the new Velocity
     
-    call velocity_from_coriolis(new_state, coriolis, lnew_velocity)    
+    call velocity_from_coriolis(new_state, coriolis, new_velocity)    
     if(dim == 3) then
       ! Recover the vertical velocity
-      ewrite_minmax(lnew_velocity%val(W_)%ptr)
-      call set(lnew_velocity, W_, new_w)
-      ewrite_minmax(lnew_velocity%val(W_)%ptr)
+      ewrite_minmax(new_velocity%val(W_)%ptr)
+      call set(new_velocity, W_, new_w)
+      ewrite_minmax(new_velocity%val(W_)%ptr)
       call deallocate(new_w)
     end if
     
     if(debug_vtus) then
-      call allocate(div, new_p_mesh, trim(lnew_velocity%name) // "Divergence")
-      call set_solver_options(temp_solver_path, ksptype = "cg", pctype = "sor", rtol = 0.0, atol = epsilon(0.0), max_its = 10000)
-      div%option_path = temp_solver_path
-      call zero(div)
-      call compute_divergence(lnew_velocity, matrices%ct_m, get_mass_matrix(new_state, new_p_mesh), div)    
-      if(gp) then
-        call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
-          & sfields = (/new_p, new_gp, div/), vfields = (/lnew_velocity, coriolis, conserv, new_res/), stat = stat)
+      if(new_velocity%mesh == new_u_mesh) then  
+        call allocate(div, new_p_mesh, trim(new_velocity%name) // "Divergence")
+        call set_solver_options(temp_solver_path, ksptype = "cg", pctype = "sor", rtol = 0.0, atol = epsilon(0.0), max_its = 10000)
+        div%option_path = temp_solver_path
+        call zero(div)
+        call compute_divergence(new_velocity, matrices%ct_m, get_mass_matrix(new_state, new_p_mesh), div)    
+        if(gp) then
+          call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
+            & sfields = (/new_p, new_gp, div/), vfields = (/new_velocity, coriolis, conserv, new_res/), stat = stat)
+        else
+          call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
+            & sfields = (/new_p, div/), vfields = (/new_velocity, coriolis, conserv, new_res/), stat = stat)
+        end if
+        if(stat /= 0) then
+          ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
+        end if
+        call delete_option(div%option_path)
+        call deallocate(div)
       else
-        call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
-          & sfields = (/new_p, div/), vfields = (/lnew_velocity, coriolis, conserv, new_res/), stat = stat)
+        if(gp) then
+          call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
+            & sfields = (/new_p, new_gp/), vfields = (/new_velocity, coriolis, conserv, new_res/), stat = stat)
+        else
+          call vtk_write_fields("geostrophic_interpolation_new", vtu_index, new_positions, model = new_u_mesh, &
+            & sfields = (/new_p/), vfields = (/new_velocity, coriolis, conserv, new_res/), stat = stat)
+        end if
+        if(stat /= 0) then
+          ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
+        end if
       end if
-      if(stat /= 0) then
-        ewrite(0, *) "WARNING: Error returned by vtk_write_fields: ", stat
-      end if
-      call delete_option(div%option_path)
-      call deallocate(div)
     end if
     call deallocate(conserv)
     
@@ -2735,8 +2761,8 @@ contains
     call deallocate(new_p)
     call deallocate(new_res)
     if(gp) call deallocate(new_gp)
-    call deallocate(lnew_velocity)
-    deallocate(lnew_velocity)
+    call deallocate(new_bc_velocity)
+    deallocate(new_bc_velocity)
     
     if(debug_vtus) then
       vtu_index = vtu_index + 1
